+++ src/config.js
@@ -5,23 +5,23 @@
  * @property {string} [site_title='Wiki'] The website title, HTML <title> format: page_title | site_title
  * @property {string} [site_header='Wiki'] Used in the navbar as your site title.
  * @property {string} [site_footer='Wiki'] Used as the footer text of your site.
- * @property {object[]} [site_sections=[]] Your site sections for homepage & tag pages. For each section, the home page & tag pages will display a section box that lists the document count for documents that have a matching tag. Clicking the section link will list the tagged documents.
+ * @property {object[]} site_sections=[] Your site sections for homepage & tag pages. For each section, the home page & tag pages will display a section box that lists the document count for documents that have a matching tag. Clicking the section link will list the tagged documents.
  * @property {string} site_sections.title Your site section header text.
  * @property {string} site_sections.description Your site section description text.
  * @property {string} site_sections.tag Your site section related tag.
  * @property {string} [home_page='home-page'] Slug of the root `/` page document.
- * @property {string[]} [ignore_slugs] Slugs to ignore in search & filtered documents, default is 'home-page';
+ * @property {string[]} ignore_slugs Slugs to ignore in search & filtered documents, default is 'home-page';
  * @property {number} [excerpt_length=400] Excerpt length, used in search result previews.
- * @property {string} [site_url=''] Application base URL. Used for canonical URLs and Open Graph, no trailing slash.
+ * @property {string} site_url='' Application base URL. Used for canonical URLs and Open Graph, no trailing slash.
  * @property {string} [theme_dir=''] Specify the path to the theme directory, no trailing slash.
  * @property {string} [public_dir=''] Path to the static file directory for themes, no trailing slash
  * @property {boolean} [use_delete_key=false] Enable hiding document deletion behind a private key.
- * @property {string} delete_key Key used for verifying document deletion.
+ * @property {string|undefined} delete_key Key used for verifying document deletion.
  * @property {boolean} [use_edit_key=false] Enable hiding document modification behind a private key.
- * @property {string} edit_key Key used for verifying document modification.
+ * @property {string|undefined} edit_key Key used for verifying document modification.
  * @property {boolean} [public_history=true] Allow access to history URLs.
  * @property {boolean} [handle_not_found=true] Allows the middleware to capture fall through routes as a `404 not found` handler when enabled.
- * @property {string[]} [allowedDocumentKeys] List of allowed custom values to set on a document. `title`, `excerpt`, `content`, `slug`, and `tags` are always allowed.
+ * @property {string[]} allowedDocumentKeys=[] List of allowed custom values to set on a document. `title`, `excerpt`, `content`, `slug`, and `tags` are always allowed.
  * @property {boolean} [use_meta_data=true] Theme specific, use OpenGraph and neta data.
  * @property {string} [site_locale='en_US'] Theme specific, Open Graph: Locale
  * @property {string} [site_twitter_site=''] Theme specific, Open Graph: Twitter Site Handle