+++ test/saveNew.test.js
@@ -174,7 +174,7 @@ test('redirects back after attempting to saving when missing body keys', async (
   const response = { redirect };
   const server = serverSetup();
   const uttori = new UttoriWiki(config, server);
-  await uttori.saveNew({ params: { slug: 'test-old', key: 'test-key' }, body: {} }, response, next);
+  await uttori.saveNew({ params: { key: 'test-key' }, body: {} }, response, next);
   t.true(redirect.calledOnce);
 });
 
@@ -186,7 +186,7 @@ test('redirects back after attempting to saving when missing body', async (t) =>
   const response = { redirect };
   const server = serverSetup();
   const uttori = new UttoriWiki(config, server);
-  await uttori.saveNew({ params: { slug: 'test-old', key: 'test-key' }, body: undefined }, response, next);
+  await uttori.saveNew({ params: { key: 'test-key' }, body: undefined }, response, next);
   t.true(redirect.calledOnce);
 });
 
@@ -198,6 +198,6 @@ test('redirects back after attempting to saving when use_edit_key is set but no 
   const response = { redirect };
   const server = serverSetup();
   const uttori = new UttoriWiki(config, server);
-  await uttori.saveNew({ params: { slug: 'test-old', key: 'bad-key' }, body: { title: 'Title' } }, response, next);
+  await uttori.saveNew({ params: { key: 'bad-key' }, body: { title: 'Title' } }, response, next);
   t.true(redirect.calledOnce);
 });