+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.0.3](https://github.com/uttori/uttori-wiki/compare/v5.0.2...v5.0.3) - 2022-11-22
+
+- 🪲 Creating new documents was not correctly checking for existing documents and showing an `undefined` flash message.
+- 🛠 Clean up some types that weren't as specific as they could be.
+- 🎁 Update dependencies
+
 ## [5.0.2](https://github.com/uttori/uttori-wiki/compare/v5.0.1...v5.0.2) - 2022-05-30
 
 - 🧰 Add `production` configuration object, useful for development.